pyproject.toml
[tool.poetry]
name = "ifdo"
version = "0.1.0"
description = "iFDO utilities"
authors = ["Kevin Barnard <kbarnard@mbari.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-benedict = "^0.28.3"
stringcase = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
